feat: Auth 支持多系统接入 #3889 格式化
