+++ src/backend/ci/core/common/common-auth/common-auth-v3/src/main/kotlin/com/tencent/devops/common/auth/api/BluekingV3AuthProjectApi.kt
@@ -38,8 +38,8 @@ import com.tencent.devops.common.auth.api.pojo.BkAuthGroupAndUserList
 import com.tencent.devops.common.auth.api.pojo.BkAuthProjectInfoResources
 import com.tencent.devops.common.auth.code.AuthServiceCode
 import com.tencent.devops.common.auth.service.IamEsbService
-import com.tencent.devops.common.auth.utlis.ActionUtils
-import com.tencent.devops.common.auth.utlis.AuthUtils
+import com.tencent.devops.common.auth.utils.ActionUtils
+import com.tencent.devops.common.auth.utils.AuthUtils
 import org.slf4j.LoggerFactory
 
 class BluekingV3AuthProjectApi constructor(
