--- src/backend/ci/core/common/common-auth/common-auth-v3/src/main/kotlin/com/tencent/devops/common/auth/utlis/ActionUtils.kt
@@ -1,15 +1,0 @@
-package com.tencent.devops.common.auth.utlis
-
-import com.tencent.devops.common.auth.api.AuthPermission
-import com.tencent.devops.common.auth.api.AuthResourceType
-
-object ActionUtils {
-
-    fun buildAction(authResourceType: AuthResourceType, permission: AuthPermission): String {
-        return if (permission == AuthPermission.LIST) {
-            "${authResourceType.value}_${AuthPermission.VIEW.value}"
-        } else {
-            "${authResourceType.value}_${permission.value}"
-        }
-    }
-}
