--- src/backend/ci/core/common/common-auth/common-auth-v3/src/main/kotlin/com/tencent/devops/common/auth/utlis/StringUtils.kt
@@ -1,28 +1,0 @@
-package com.tencent.devops.common.auth.utlis
-
-import java.nio.charset.Charset
-import java.util.Base64
-
-object StringUtils {
-    fun obj2List(str: String): List<String> {
-        val list = str.substringBefore("]").substringAfter("[").split(",")
-        val newList = mutableListOf<String>()
-        list.map {
-            newList.add(it.trim())
-        }
-        return newList
-    }
-
-    fun removeAllElement(set: Set<String>): Set<String> {
-        if (set.contains("*")) {
-            val newSet = mutableSetOf<String>()
-            set.map {
-                if (it != "*") {
-                    newSet.add(it)
-                }
-                return newSet
-            }
-        }
-        return set
-    }
-}
